Refactor PatchFinder class to decouple responsibility
- Created different Ranger classes to handle different sources of information
