+++ src/main/java/eu/fasten/vulnerabilityproducer/utils/connections/GitHubAPI.java
@@ -18,7 +18,7 @@
 
 package eu.fasten.vulnerabilityproducer.utils.connections;
 
-import eu.fasten.vulnerabilityproducer.utils.PatchFarmer;
+import eu.fasten.vulnerabilityproducer.utils.PatchFinder;
 import org.jooq.tools.json.JSONArray;
 import org.jooq.tools.json.JSONObject;
 import org.jooq.tools.json.JSONParser;
@@ -255,7 +255,7 @@ public class GitHubAPI {
                 var node = (JSONObject) nodeObj;
                 if (node.get("body") != null) {
                     // Grab potential links
-                    var links = PatchFarmer.findPatchLinks(node.get("body").toString());
+                    var links = PatchFinder.findPatchLinks(node.get("body").toString());
                     for (String ref : links) {
                         if (ref.contains(repoOwner + "/" + repoName)) {
                             refLinks.add(ref);
