+++ src/main/java/eu/fasten/vulnerabilityproducer/utils/parsers/ParserManager.java
@@ -20,14 +20,17 @@ package eu.fasten.vulnerabilityproducer.utils.parsers;
 
 import com.mongodb.client.MongoDatabase;
 import eu.fasten.vulnerabilityproducer.db.NitriteController;
-import eu.fasten.vulnerabilityproducer.utils.PatchFarmer;
+import eu.fasten.vulnerabilityproducer.utils.PatchFinder;
 import eu.fasten.vulnerabilityproducer.utils.Vulnerability;
 import eu.fasten.vulnerabilityproducer.utils.connections.JavaHttpClient;
 import eu.fasten.vulnerabilityproducer.utils.mappers.VersionRanger;
+import org.apache.commons.io.FileUtils;
 import org.jooq.tools.json.JSONParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 
@@ -42,7 +45,7 @@ public class ParserManager {
     private String ghToken;
     private NVDParser nvdParser;
     private OVALParser ovalParser;
-    private PatchFarmer patchFarmer;
+    private PatchFinder patchFinder;
     private VersionRanger versionRanger;
     private NitriteController nitriteController;
     private final Logger logger = LoggerFactory.getLogger(ParserManager.class.getName());
@@ -59,7 +62,7 @@ public class ParserManager {
         this.ovalParser = new OVALParser(client, versionRanger);
         this.ghParser = new GHParser(client, ghToken, versionRanger, pathToGHCursor);
         this.nvdParser = new NVDParser(new JSONParser(), client);
-        this.patchFarmer = new PatchFarmer(mongoDatabase, client);
+        this.patchFinder = new PatchFinder(mongoDatabase, client);
         this.nitriteController = nitriteController;
     }
 
@@ -93,7 +96,15 @@ public class ParserManager {
 
         logger.info("Injecting information about Patches");
         for (Vulnerability v : vulnerabilities) {
-            patchFarmer.parseReferences(v, nitriteController);
+            // Adding patch information
+            patchFinder.parseReferences(v, nitriteController);
+            // Storing the file to statements
+            File file = new File("/Users/Edo/Desktop/statements/" + v.getId() + ".json");
+            try {
+                FileUtils.writeStringToFile(file, v.toJson());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
 
         filterNewEntries(vulnerabilities);
@@ -132,7 +143,7 @@ public class ParserManager {
         logger.info("Parsing references of all vulnerabilities to inject patches information");
         for (Vulnerability v : vulnerabilities) {
             logger.info("Parsing references from Vulnerability with ID: " + v.getId());
-            patchFarmer.parseReferences(v, nitriteController);
+            patchFinder.parseReferences(v, nitriteController);
         }
 
         logger.info("Filtering if entries are new using Nitrite Controller");
@@ -200,17 +211,17 @@ public class ParserManager {
      * @param nvdParser   - NVDParser
      * @param ghParser    - GHParser
      * @param extraParser - ExtraParser
-     * @param patchFarmer - PatchFarmer
+     * @param patchFinder - PatchFarmer
      */
     public void injectParsers(NVDParser nvdParser,
                               GHParser ghParser,
                               ExtraParser extraParser,
                               OVALParser ovalParser,
-                              PatchFarmer patchFarmer) {
+                              PatchFinder patchFinder) {
         this.nvdParser = nvdParser;
         this.ghParser = ghParser;
         this.extraParser = extraParser;
         this.ovalParser = ovalParser;
-        this.patchFarmer = patchFarmer;
+        this.patchFinder = patchFinder;
     }
 }
