+++ src/main/java/eu/fasten/vulnerabilityproducer/utils/patches/GitLabRanger.java
@@ -1,0 +1,4 @@
+package eu.fasten.vulnerabilityproducer.utils.patches;
+
+public class GitLabRanger {
+}
