+++ src/test/java/eu/fasten/vulnerabilityproducer/PatchFinderTest.java
@@ -22,8 +22,7 @@ import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import eu.fasten.vulnerabilityproducer.db.NitriteController;
-import eu.fasten.vulnerabilityproducer.db.PatchObject;
-import eu.fasten.vulnerabilityproducer.utils.PatchFarmer;
+import eu.fasten.vulnerabilityproducer.utils.PatchFinder;
 import eu.fasten.vulnerabilityproducer.utils.connections.JavaHttpClient;
 import eu.fasten.vulnerabilityproducer.utils.Vulnerability;
 import org.apache.commons.io.FileUtils;
@@ -44,7 +43,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.when;
 
-public class PatchFarmerTest {
+public class PatchFinderTest {
 
     // Inject Java Http Client to test
     JavaHttpClient clientMock = Mockito.mock(JavaHttpClient.class);
@@ -55,7 +54,7 @@ public class PatchFarmerTest {
     MongoCollection<Document> prsMock = Mockito.mock(MongoCollection.class);
     MongoCollection<Document> issuesMock = Mockito.mock(MongoCollection.class);
     NitriteController nc = Mockito.mock(NitriteController.class);
-    private PatchFarmer patchFarmer;
+    private PatchFinder patchFinder;
 
     static final String validJSON = "{" +
             "\"comment\":\"Testing\"," +
@@ -164,7 +163,7 @@ public class PatchFarmerTest {
         when(dbMock.getCollection("commits")).thenReturn(commitsMock);
         when(dbMock.getCollection("pull_requests")).thenReturn(prsMock);
         when(dbMock.getCollection("issues")).thenReturn(issuesMock);
-        patchFarmer = new PatchFarmer(dbMock, clientMock);
+        patchFinder = new PatchFinder(dbMock, clientMock);
     }
 
     public String readFile(String filePath) throws IOException {
@@ -185,7 +184,7 @@ public class PatchFarmerTest {
         when(mockDocument.toJson()).thenReturn(jsonCommit);
 
         // Call parse Commit method
-        var vp = patchFarmer.parseGHCommit(patchLink);
+        var vp = patchFinder.parseReference(patchLink);
 
         // Verify Hunks in the commit
         List<Integer> hunks = new ArrayList<>();
@@ -216,7 +215,7 @@ public class PatchFarmerTest {
 
         // Call parse Commit method
         v.addReference(patchLink);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         // Verify Hunks in the commit
         List<Integer> hunks = new ArrayList<>();
@@ -251,7 +250,7 @@ public class PatchFarmerTest {
         when(mockDocumentMergeCommit.toJson()).thenReturn(jsonMergeCommit);
 
         // Call parse Commit method
-        var vp = patchFarmer.parseGHPullRequest(prLink);
+        var vp = patchFinder.parseReference(prLink);
 
         // Verify Hunks in the commit
         List<Integer> hunks = new ArrayList<>();
@@ -297,7 +296,7 @@ public class PatchFarmerTest {
 
         // Call parse Commit method
         v.addReference(prLink);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         // Verify Hunks in the commit
         List<Integer> hunks = new ArrayList<>();
@@ -346,7 +345,7 @@ public class PatchFarmerTest {
         when(mockDocumentMergeCommit.toJson()).thenReturn(jsonMergeCommit);
 
         // Call parse Commit method
-        var vp = patchFarmer.parseGHIssue(issueLink);
+        var vp = patchFinder.parseReference(issueLink);
 
         // Verify Hunks in the commit
         List<Integer> hunks = new ArrayList<>();
@@ -401,7 +400,7 @@ public class PatchFarmerTest {
 
         // Call parse Commit method
         v.addReference(issueLink);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         // Verify Hunks in the commit
         List<Integer> hunks = new ArrayList<>();
@@ -430,7 +429,7 @@ public class PatchFarmerTest {
         files.add(f);
         when(nc.findPatchEntry("ref")).thenReturn(Optional.of(files));
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
         assertTrue(v.getPatches().contains(f));
     }
 
@@ -455,7 +454,7 @@ public class PatchFarmerTest {
         List<Integer> ln_2 = new ArrayList<>(Arrays.asList(2194, 2204, 2211));
         patch_2.setLineNumbers(ln_2);
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertTrue(v.getPatches().contains(patch_1));
         assertTrue(v.getPatches().contains(patch_2));
@@ -482,7 +481,7 @@ public class PatchFarmerTest {
         patch_1.setLineNumbers(ln_1);
         patch_1.setPatchDate("2010-04-21");
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertTrue(v.getPatches().contains(patch_1));
     }
@@ -515,7 +514,7 @@ public class PatchFarmerTest {
         List<Integer> ln_2 = new ArrayList<>(Arrays.asList(327));
         patch_2.setLineNumbers(ln_2);
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertTrue(v.getPatches().contains(patch_1));
         assertTrue(v.getPatches().contains(patch_2));
@@ -550,7 +549,7 @@ public class PatchFarmerTest {
         Vulnerability v = new Vulnerability("TEST");
         v.addReference(refLink);
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertTrue(v.getPatches().contains(patch));
     }
@@ -571,7 +570,7 @@ public class PatchFarmerTest {
         Vulnerability v = new Vulnerability("TEST");
         v.addReference(ref);
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
         assertTrue(v.getPatches().contains(patch));
     }
 
@@ -607,7 +606,7 @@ public class PatchFarmerTest {
         List<Integer> ln_2 = new ArrayList<>(Arrays.asList(327));
         patch_2.setLineNumbers(ln_2);
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertTrue(v.getPatches().contains(patch_1));
         assertTrue(v.getPatches().contains(patch_2));
@@ -646,7 +645,7 @@ public class PatchFarmerTest {
 
         Vulnerability v = new Vulnerability();
         v.addPatchLink(ref);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertEquals(patches_exp, v.getPatches());
     }
@@ -678,7 +677,7 @@ public class PatchFarmerTest {
 
         Vulnerability v = new Vulnerability();
         v.addPatchLink(ref);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertEquals(patches_exp, v.getPatches());
     }
@@ -712,7 +711,7 @@ public class PatchFarmerTest {
 
         Vulnerability v = new Vulnerability();
         v.addPatchLink(ref);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertEquals(patches_exp, v.getPatches());
     }
@@ -744,7 +743,7 @@ public class PatchFarmerTest {
 
         Vulnerability v = new Vulnerability();
         v.addPatchLink(ref);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertEquals(patches_exp, v.getPatches());
     }
@@ -788,7 +787,7 @@ public class PatchFarmerTest {
         patch_3.setLineNumbers(ln_3);
         patch_3.setPatchDate("2020-06-03");
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertTrue(v.getPatches().contains(patch_1));
         assertTrue(v.getPatches().contains(patch_2));
@@ -838,7 +837,7 @@ public class PatchFarmerTest {
         patch_3.setLineNumbers(ln_3);
         patch_3.setPatchDate("2018-06-12");
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertTrue(v.getPatches().contains(patch_1));
         assertTrue(v.getPatches().contains(patch_2));
@@ -875,7 +874,7 @@ public class PatchFarmerTest {
         patch_1.setLineNumbers(ln_1);
         patch_1.setPatchDate("2019-07-04");
 
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertTrue(v.getPatches().contains(patch_1));
     }
@@ -902,7 +901,7 @@ public class PatchFarmerTest {
 
         Vulnerability v = new Vulnerability();
         v.addPatchLink(refLink);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertEquals(patches_exp, v.getPatches());
     }
@@ -932,7 +931,7 @@ public class PatchFarmerTest {
 
         Vulnerability v = new Vulnerability();
         v.addPatchLink(refLink);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertEquals(patches_exp, v.getPatches());
     }
@@ -965,7 +964,7 @@ public class PatchFarmerTest {
 
         Vulnerability v = new Vulnerability();
         v.addPatchLink(refLink);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertEquals(patches_exp, v.getPatches());
     }
@@ -995,7 +994,7 @@ public class PatchFarmerTest {
 
         Vulnerability v = new Vulnerability();
         v.addPatchLink(refLink);
-        patchFarmer.parseReferences(v, nc);
+        patchFinder.parseReferences(v, nc);
 
         assertEquals(patches_exp, v.getPatches());
     }
