+++ src/test/java/eu/fasten/vulnerabilityproducer/parsers/ParserManagerTest.java
@@ -20,7 +20,7 @@ package eu.fasten.vulnerabilityproducer.parsers;
 
 import com.mongodb.client.MongoDatabase;
 import eu.fasten.vulnerabilityproducer.db.NitriteController;
-import eu.fasten.vulnerabilityproducer.utils.PatchFarmer;
+import eu.fasten.vulnerabilityproducer.utils.PatchFinder;
 import eu.fasten.vulnerabilityproducer.utils.Vulnerability;
 import eu.fasten.vulnerabilityproducer.utils.connections.JavaHttpClient;
 import eu.fasten.vulnerabilityproducer.utils.mappers.Severity;
@@ -44,7 +44,7 @@ public class ParserManagerTest {
     NVDParser nvdParserMock = Mockito.mock(NVDParser.class);
     OVALParser ovalParserMock = Mockito.mock(OVALParser.class);
     ExtraParser extraParserMock = Mockito.mock(ExtraParser.class);
-    PatchFarmer patchFarmerMock = Mockito.mock(PatchFarmer.class);
+    PatchFinder patchFinderMock = Mockito.mock(PatchFinder.class);
 
     @Test
     public void testGetDataAndMergeFromParsers() {
@@ -53,7 +53,7 @@ public class ParserManagerTest {
                 ghParserMock,
                 extraParserMock,
                 ovalParserMock,
-                patchFarmerMock);
+                patchFinderMock);
 
         HashMap<String, Vulnerability> nvdVulns = new HashMap<>();
         Vulnerability v1 = new Vulnerability("CVE-TEST-1");
@@ -138,7 +138,7 @@ public class ParserManagerTest {
                 ghParserMock,
                 extraParserMock,
                 ovalParserMock,
-                patchFarmerMock);
+                patchFinderMock);
 
         HashMap<String, Vulnerability> nvdVulns = new HashMap<>();
         Vulnerability v1 = new Vulnerability("CVE-TEST-1");
